+++ CHANGELOG.md
@@ -25,6 +25,9 @@ Unless otherwise specified, any version comparison below is the comparison of se
   overloads for this method are also added.
 - Block breaking particle won't be sent if block is broken by flowing liquid.
 - Water placed in nether dimension will disappear immediately now.
+- `Pos`, `Motion` and `Rotation` in entity nbt are now saved as list tag instead of compound tag to match vanilla,
+  this also fixed the bug that entities being spawned in the right position when placing structure using `/structure`
+  command. Please note that this change is not backward compatible and will break the old world and player data.
 
 ## 0.1.1 (API 0.2.0)